pyproject.toml
[project]
name = "amon-claw"
version = "0.0.1"
description = "Meu Simples OpenClaw, feito apenas para me atender"
authors = [
    { email = "matheus.amon@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.1.0",
    "fastapi[standard]>=0.135.3",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-redis>=0.4.0",
    "langgraph-store-mongodb>=0.2.0",
    "motor>=3.7.1",
    "pydantic-ai-slim[openrouter]>=1.70.0",
    "pydantic-settings>=2.13.1",
]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 90

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["amon_claw"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[dependency-groups]
dev = [
    "mongomock-motor>=0.0.36",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.7",
]
